requirements.txt
streamlit
PyPDF2
langchain
langchain_community
langchain_cohere
cohere
faiss-cpu
python-docx
pymupdf
pillow